chat_django/requirements.txt
django>=5.0.0
anthropic>=0.18.1
openai
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
redis>=5.0.1
dataclasses-json>=0.6.4
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
mcp>=1.2.0
httpx>=0.26.0

selenium
mod_proxy_wss